pyproject.toml
[tool.poetry]
name = "pydff-matrix"
version = "0.1.0"
description = ""
authors = ["JensDiemer <git@jensdiemer.de>"]

[tool.poetry.dependencies]
python = "^3.6"
matrix-synapse = "*"
certbot = "*"
python-env-loader = "*"

# https://github.com/poljar/matrix-nio
# nio without end-to-end encryption support:
matrix-nio = "*"

# For e2ee support, python-olm is needed which requires the libolm C library (version 3.x):
#matrix-nio = { version = "*", extras = ["e2e"] }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
